refactor: extract getVisiblePages logic into isolated utility with unit tests
